pyproject.toml
[project]
name = "pola"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["rio-ui==0.10.4"]
